Script to modify block styles
